% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompDB.R
\name{CompDB}
\alias{CompDB}
\title{Trajectories comparison}
\usage{
CompDB(data, axis, nref, ncomp)
}
\arguments{
\item{data}{is the data base}

\item{axis}{are the groups or levels which are defined to compare across the trajectory}

\item{nref}{is the column or the set of columns of reference which you are interested to compare}

\item{ncomp}{is the column or set of clumns used to compare "ref"}
}
\description{
The function is useful to compare the means between set of groups
}
\details{
This function shows a plot with the trajectory comparison between "ref" and "comp" arguments
}
\examples{
month=1:12 ; A=trunc(rnorm(12,20,2)) ; B=trunc(rnorm(12,20,10)) ; C=trunc(rnorm(12,20,12))
D=trunc(rnorm(12,20,15)) ; df=data.frame(month,A,B,C,D)
example 1
CompDB(data=df,axis="month",nref="A",ncomp=c("B","C"))
example 2
CompDB(data=df,axis="month",nref=c("A","B"),ncomp=c("C","D"))
}
